capture-app/zed/pyproject.toml
[project]
name = "zed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "common",
    "pyzed==5.1 ; ((sys_platform == 'linux' and platform_machine == 'aarch64') or (sys_platform == 'win32' and platform_machine == 'AMD64')) and platform_python_implementation == 'CPython' and python_version >= '3.13' and python_version < '3.14'",
    "fastapi>=0.116.1",
    "scipy>=1.16.1",
    "uvicorn>=0.35.0",
    "openapi-generator-cli>=7.14.0",
    "typer>=0.20.0",
    "pillow>=12.0.0",
]

[dependency-groups]
dev = ["scipy-stubs"]

[tool.uv.sources]
common = { path = "../../common", editable = true }
pyzed = { index = "pyzed" }

[[tool.uv.index]]
name = "pyzed"
url = "third-party/pyzed"
format = "flat"
explicit = true

[tool.ruff]
extend = "../../ruff.toml"
